+++ public/vstdlib/vstrtools.h
@@ -27,7 +27,7 @@
 // conversion functions wchar_t <-> char, returning the number of characters
 // converted
 VSTRTOOLS_INTERFACE int V_UTF8ToUnicode(const char *pUTF8, wchar_t *pwchDest,
-                                        int cubDestSizeInBytes);
+                                        intp cubDestSizeInBytes);
 VSTRTOOLS_INTERFACE int V_UnicodeToUTF8(const wchar_t *pUnicode, char *pUTF8,
                                         int cubDestSizeInBytes);
 VSTRTOOLS_INTERFACE intp V_UCS2ToUnicode(const ucs2 *pUCS2, wchar_t *pUnicode,
